ironrdp_client: remove dependence on Rustls for connection sequence
